+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
-[no unreleased changes yet]
+- **BREAKING:** Rename `MemoizingMethods` config option for `RungerStyle/FirstArgumentIndentation` to `DecoratorMethods`.
+- Add `helper_method` to default `RungerStyle/FirstArgumentIndentation` `DecoratorMethods`.
+- Add `RungerStyle/ArgumentAlignment` cop, which ignores [macro](https://rubydoc.info/github/rubocop/rubocop-ast/RuboCop/AST/MethodDispatchNode#macro%3F-instance_method) methods.
+- Remove `Layout/MultilineAssignmentLayout` custom config that included `send`.
+- Specify `EnforcedStyle: consistent` for `Layout/FirstHashElementIndentation`.
 
 ## v4.5.0 (2025-02-11)
 - Add custom `RungerStyle/MultilineHashValueIndentation` cop.